requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
playwright>=1.40.0
aiohttp>=3.9.0
fake-useragent>=1.4.0
python-dotenv>=1.0.0
pyyaml>=6.0
pandas>=2.1.0
openpyxl>=3.1.0
